requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
Django==5.1.5
django-cors-headers==4.6.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
Faker==33.3.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
injector==0.22.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
pycparser==2.22
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.7.0